+++ KeyboardKitDemoKeyboard/Demo/DemoKeyboardActionHandler.swift
@@ -56,20 +56,16 @@ class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
     func saveImage(_ image: UIImage) {
         guard let input = inputViewController else { return }
         guard input.hasFullAccess else { return alert("You must enable full access to save images.") }
-        let saveCompletion = #selector(handleImage(_:didFinishSavingWithError:contextInfo:))
-        image.saveToPhotos(completionTarget: self, completionSelector: saveCompletion)
-    }
-
-
-    // MARK: - Image Functions
-
-    @objc func handleImage(_ image: UIImage, didFinishSavingWithError error: Error?, contextInfo: UnsafeRawPointer) {
-        if error == nil { alert("Saved!") }
-        else { alert("Failed!") }
+        image.saveToPhotos(completion: handleImageDidSave)
     }
 }
 
 private extension DemoKeyboardActionHandler {
+    
+    func handleImageDidSave(WithError error: Error?) {
+        if error == nil { alert("Saved!") }
+        else { alert("Failed!") }
+    }
     
     func switchToEmojiKeyboardCategory(_ cat: EmojiCategory) {
         guard
@@ -77,34 +73,5 @@ private extension DemoKeyboardActionHandler {
             let view = vc.emojiCollectionView
             else { return }
         view.moveToSection(byCategory: cat.title)
-    }
-}
-
-private extension UIImage {
-    
-    func saveToPhotos(completionTarget: AnyObject?, completionSelector: Selector?) {
-        UIImageWriteToSavedPhotosAlbum(self, completionTarget, completionSelector, nil)
-    }
-    
-    func saveToPhotos(completion: @escaping (Error?) -> Void) {
-        let service = StandardPhotosImageService()
-        service.saveImageToPhotos(self, completion: completion)
-    }
-}
-
-private class StandardPhotosImageService: NSObject {
-    
-    public typealias Completion = (Error?) -> Void
-
-    private var completions = [Completion]()
-    
-    public func saveImageToPhotos(_ image: UIImage, completion: @escaping (Error?) -> Void) {
-        completions.append(completion)
-        image.saveToPhotos(completionTarget: self, completionSelector: #selector(didSave(_:error:contextInfo:)))
-    }
-    
-    @objc func didSave(_ image: UIImage, error: NSError?, contextInfo: UnsafeRawPointer) {
-        guard completions.count > 0 else { return }
-        completions.removeFirst()(error)
     }
 }